Split out tree prior implementation
Transfer most places where trees or tree priors are used to a
separate sub-package (without affecting the implementation) and
mark all other spots by references to that sub-package.

This is the first step towards implementing a more flexible and
transparent handling of tree priors, which will hopefully culminate
in an implementation of reticulate phylogeny inference.